Update build configuration and documentation setup
Changed the build environment in `dotnet.yml` from `ubuntu-latest` to `windows-latest`. Added a step to restore the `Documentation/Documentation.shfbproj` project and introduced a job to build the documentation with a `Release` configuration.

In `Documentation.shfbproj`, added `<ComponentPath>` and `<SHFBROOT>` properties for package management. Included new package references for `EWSoftware.SHFB` and `EWSoftware.SHFB.NET` with specified versions.